+++ packages/frontend/src/components/TableCalculationModels/index.tsx
@@ -1,0 +1,3 @@
+export { default as CreateTableCalculationModal } from './CreateTableCalculationModal';
+export { default as DeleteTableCalculationModal } from './DeleteTableCalculationModal';
+export { default as UpdateTableCalculationModal } from './UpdateTableCalculationModal';